Move the function 'check_resize_osd_pre_conditions' to the 'resize_osd' file, Create a new function 'update_resize_osd_count'
